+++ client/Packages/com.beamable.server/Editor/MicroserviceEditor.cs
@@ -87,6 +87,7 @@ namespace Beamable.Server.Editor
 					return;
 				}
 
+				
 				TryToPreloadBaseImage();
 				TryToPreloadMongoImage();