+++ client/Packages/com.beamable/Editor/BeamEditor.cs
@@ -475,6 +475,8 @@ namespace Beamable
 		public Action OnServiceArchived;
 		public Action OnServiceUnarchived;
 
+		public OptionalString RealmSecret { get; private set; } = new OptionalString();
+
 		public void Init(string playerCode, IDependencyBuilder builder)
 		{
 			PlayerCode = playerCode;
@@ -551,6 +553,8 @@ namespace Beamable
 #pragma warning restore CS4014
 						await Promise.Success;
 					}
+
+					await ResetRealmSecret();
 				}
 			}
 
@@ -933,7 +937,6 @@ namespace Beamable
 		public Promise<string> GetRealmSecret()
 		{
 			// TODO this will only work if the current user is an admin.
-
 			return Requester.Request<CustomerResponse>(Method.GET, "/basic/realms/admin/customer").Map(resp =>
 			{
 				var matchingProject = resp.customer.projects.FirstOrDefault(p => p.name.Equals(CurrentRealm.Pid));
@@ -961,6 +964,22 @@ namespace Beamable
 		{
 			return SwitchRealm(realm.FindRoot(), realm?.Pid);
 		}
+
+		public async Promise<OptionalString> ResetRealmSecret()
+		{
+			try
+			{
+				var secret = await GetRealmSecret();
+				RealmSecret.SetValue(secret);
+			}
+			catch
+			{
+				// this is expected to fail if the user doesn't have permission to get the realm secret.
+				RealmSecret.Clear();
+			}
+			return RealmSecret;
+		}
+
 
 		public async Promise SwitchRealm(RealmView game, string pid)
 		{
@@ -991,6 +1010,7 @@ namespace Beamable
 				CurrentRealm = realm;
 				await SaveRealmInConfig();
 				await ServiceScope.GetService<ContentIO>().FetchManifest();
+				await ResetRealmSecret();
 				OnRealmChange?.Invoke(realm);
 				ProductionRealm = game;
 				return;
@@ -999,9 +1019,9 @@ namespace Beamable
 			{
 				await ServiceScope.GetService<ContentIO>().FetchManifest();
 			}
-
 			ProductionRealm = game;
 			await SaveRealmInConfig();
+			await ResetRealmSecret();
 		}
 
 		private async Promise SaveRealmInConfig()